+++ instantsearch-android-core/src/main/kotlin/com/algolia/instantsearch/core/relevantsort/RelevantSortPriority.kt
@@ -1,11 +1,11 @@
-package com.algolia.instantsearch.core.smartsort
+package com.algolia.instantsearch.core.relevantsort
 
 /**
  * Represents the priority to apply to the search in the dynamically sorted index
  *
  * @param relevancyStrictness relevancy strictness value to apply to the search
  */
-public enum class SmartSortPriority(
+public enum class RelevantSortPriority(
     public val relevancyStrictness: Int
 ) {
 
@@ -18,7 +18,7 @@ public enum class SmartSortPriority(
     public companion object {
 
         @JvmStatic
-        public fun of(relevancyStrictness: Int): SmartSortPriority = when (relevancyStrictness) {
+        public fun of(relevancyStrictness: Int): RelevantSortPriority = when (relevancyStrictness) {
             0 -> HitsCount
             else -> Relevancy
         }