+++ instantsearch-android-core/src/main/kotlin/com/algolia/instantsearch/core/smartsort/SmartSortPresenter.kt
@@ -1,6 +1,0 @@
-package com.algolia.instantsearch.core.smartsort
-
-/**
- * Defines the way we want to interact with a Smart sort priority value.
- */
-public typealias SmartSortPresenter<T> = (SmartSortPriority?) -> T