+++ instantsearch-android/src/main/java/com/algolia/instantsearch/helper/relevantsort/internal/RelevantSortConnectionSingleIndex.kt
@@ -1,21 +1,21 @@
-package com.algolia.instantsearch.helper.smartsort.internal
+package com.algolia.instantsearch.helper.relevantsort.internal
 
 import com.algolia.instantsearch.core.Callback
 import com.algolia.instantsearch.core.connection.ConnectionImpl
-import com.algolia.instantsearch.core.smartsort.SmartSortPriority
-import com.algolia.instantsearch.core.smartsort.SmartSortViewModel
+import com.algolia.instantsearch.core.relevantsort.RelevantSortPriority
+import com.algolia.instantsearch.core.relevantsort.RelevantSortViewModel
 import com.algolia.instantsearch.helper.searcher.SearcherSingleIndex
 import com.algolia.search.model.response.ResponseSearch
 
 /**
- * Connection between smart sort's view model and a single index searcher.
+ * Connection between relevant sort's view model and a single index searcher.
  */
-internal class SmartSortConnectionSingleIndex(
-    val viewModel: SmartSortViewModel,
+internal class RelevantSortConnectionSingleIndex(
+    val viewModel: RelevantSortViewModel,
     val searcher: SearcherSingleIndex,
 ) : ConnectionImpl() {
 
-    private val priorityCallback: Callback<SmartSortPriority?> = callback@{ priority ->
+    private val priorityCallback: Callback<RelevantSortPriority?> = callback@{ priority ->
         if (priority == null) return@callback
         searcher.query.relevancyStrictness = priority.relevancyStrictness
         searcher.searchAsync()
@@ -27,7 +27,7 @@ internal class SmartSortConnectionSingleIndex(
             viewModel.priority.value = null
             return@callback
         }
-        val dynamicSortPriority = SmartSortPriority.of(receivedRelevancyStrictness)
+        val dynamicSortPriority = RelevantSortPriority.of(receivedRelevancyStrictness)
         if (dynamicSortPriority != viewModel.priority.value) {
             viewModel.priority.value = dynamicSortPriority
         }