+++ instantsearch-android/src/main/java/com/algolia/instantsearch/helper/relevantsort/internal/RelevantSortConnectorMultipleIndex.kt
@@ -1,16 +1,16 @@
-package com.algolia.instantsearch.helper.smartsort.internal
+package com.algolia.instantsearch.helper.relevantsort.internal
 
 import com.algolia.instantsearch.core.connection.ConnectionImpl
-import com.algolia.instantsearch.core.smartsort.SmartSortConnector
-import com.algolia.instantsearch.core.smartsort.SmartSortViewModel
+import com.algolia.instantsearch.core.relevantsort.RelevantSortConnector
+import com.algolia.instantsearch.core.relevantsort.RelevantSortViewModel
 import com.algolia.instantsearch.helper.searcher.SearcherMultipleIndex
-import com.algolia.instantsearch.helper.smartsort.connectSearcher
+import com.algolia.instantsearch.helper.relevantsort.connectSearcher
 
-internal class SmartSortConnectorMultipleIndex(
+internal class RelevantSortConnectorMultipleIndex(
     override val searcher: SearcherMultipleIndex,
     queryIndex: Int,
-    override val viewModel: SmartSortViewModel = SmartSortViewModel(),
-) : ConnectionImpl(), SmartSortConnector {
+    override val viewModel: RelevantSortViewModel = RelevantSortViewModel(),
+) : ConnectionImpl(), RelevantSortConnector {
 
     private val connectionSearcher = viewModel.connectSearcher(searcher, queryIndex)